py_backend/requirements.txt
web3==5.31.3
py-solc-x==1.1.1
python-dotenv==0.21.1
flask==2.2.3